requirements.txt
amqp==1.4.9
ansible==2.0.0.2
anyjson==0.3.3
argparse==1.2.1
astroid==1.4.3
bcrypt==2.0.0
billiard==3.3.0.22
boto==2.38.0
celery==3.1.19
cffi==1.3.0
colorama==0.3.5
coverage==4.0.3
coveralls==1.1
decorator==4.0.4
Django==1.9.1
django-autologin==0.0.3
django-celery==3.1.17
django-email-from-template==2.0.0
django-enumfield==1.2.1
django-setuptest==0.2
django-staticfiles-dotd==1.1.0
django-storages==1.1.8
django-yadt==2.0.3
ecdsa==0.13
funcsigs==0.4
gunicorn==19.3.0
ipython==4.0.0
ipython-genutils==0.1.0
Jinja2==2.8
kombu==3.0.33
lxml==3.4.4
MarkupSafe==0.23
mccabe==0.3.1
monkeypatch==0.1rc3
mutagen==1.31
paramiko==1.16.0
path.py==8.1.2
pbr==1.8.1
pep257==0.7.0
pep8==1.6.2
pep8-naming==0.3.3
pexpect==4.0.1
pickleshare==0.5
pil-compat==1.0.0
Pillow==3.0.0
progressbar==2.3
psycopg2==2.6.1
ptyprocess==0.5
pycparser==2.14
pycrypto==2.6.1
pyflakes==1.0.0
pylint==1.5.2
pylint-flask==0.1
python-memcached==1.57
pytz==2015.7
PyYAML==3.11
redis==2.10.5
requests==2.8.1
roman==2.0.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
Werkzeug==0.11.2
wsgiref==0.1.2
